Fix Code Factor warnings
